move getattr to snappy (not used anywhere else)
